requirements.txt
Shapely==1.8.0
scipy==1.8.1
diskcache==5.2.1
moviepy==1.0.3
torch==1.12.1+cu116
pytorch_lightning==1.6.0
tqdm==4.64.0
open3d==0.15.2
numpy==1.17.4
matplotlib==3.4.2
Click==7.0
PyYAML==6.0
